+++ governance/checkpoint.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"code.vegaprotocol.io/protos/vega"
-	checkpoint "code.vegaprotocol.io/protos/vega/checkpoint/v1"
+	checkpointpb "code.vegaprotocol.io/protos/vega/checkpoint/v1"
 	"code.vegaprotocol.io/vega/events"
 	"code.vegaprotocol.io/vega/types"
 
@@ -19,14 +19,14 @@ func (e *Engine) Checkpoint() ([]byte, error) {
 	if len(e.enactedProposals) == 0 {
 		return nil, nil
 	}
-	snap := &checkpoint.Proposals{
+	snap := &checkpointpb.Proposals{
 		Proposals: e.getCheckpointProposals(),
 	}
 	return proto.Marshal(snap)
 }
 
 func (e *Engine) Load(ctx context.Context, data []byte) error {
-	snap := &checkpoint.Proposals{}
+	snap := &checkpointpb.Proposals{}
 	if err := proto.Unmarshal(data, snap); err != nil {
 		return err
 	}
@@ -44,7 +44,7 @@ func (e *Engine) Load(ctx context.Context, data []byte) error {
 			Proposal: prop,
 		})
 	}
-	// sned events for restored proposals
+	// send events for restored proposals
 	e.broker.SendBatch(evts)
 	// @TODO ensure OnChainTimeUpdate is called
 	return nil