+++ types/governance.go
@@ -184,12 +184,15 @@ type ProposalSubmission struct {
 	Reference string
 	// Proposal configuration and the actual change that is meant to be executed when proposal is enacted
 	Terms *ProposalTerms
+	// Rationale behind the proposal change.
+	Rationale *ProposalRationale
 }
 
 func ProposalSubmissionFromProposal(p *Proposal) *ProposalSubmission {
 	return &ProposalSubmission{
 		Reference: p.Reference,
 		Terms:     p.Terms,
+		Rationale: p.Rationale,
 	}
 }
 
@@ -201,6 +204,11 @@ func NewProposalSubmissionFromProto(p *commandspb.ProposalSubmission) *ProposalS
 	return &ProposalSubmission{
 		Reference: p.Reference,
 		Terms:     pterms,
+		Rationale: &ProposalRationale{
+			Description: p.Rationale.Description,
+			Hash:        p.Rationale.Hash,
+			URL:         p.Rationale.Url,
+		},
 	}
 }
 
@@ -212,6 +220,11 @@ func (p ProposalSubmission) IntoProto() *commandspb.ProposalSubmission {
 	return &commandspb.ProposalSubmission{
 		Reference: p.Reference,
 		Terms:     terms,
+		Rationale: &vegapb.ProposalRationale{
+			Description: p.Rationale.Description,
+			Hash:        p.Rationale.Hash,
+			Url:         p.Rationale.URL,
+		},
 	}
 }
 
@@ -222,6 +235,7 @@ type Proposal struct {
 	State        ProposalState
 	Timestamp    int64
 	Terms        *ProposalTerms
+	Rationale    *ProposalRationale
 	Reason       ProposalError
 	ErrorDetails string
 }
@@ -290,6 +304,7 @@ func ProposalFromProto(pp *vegapb.Proposal) *Proposal {
 		Timestamp:    pp.Timestamp,
 		Terms:        ProposalTermsFromProto(pp.Terms),
 		Reason:       pp.Reason,
+		Rationale:    ProposalRationaleFromProto(pp.Rationale),
 		ErrorDetails: pp.ErrorDetails,
 	}
 }
@@ -299,7 +314,7 @@ func (p Proposal) IntoProto() *vegapb.Proposal {
 	if p.Terms != nil {
 		terms = p.Terms.IntoProto()
 	}
-	return &vegapb.Proposal{
+	proposal := &vegapb.Proposal{
 		Id:           p.ID,
 		Reference:    p.Reference,
 		PartyId:      p.Party,
@@ -309,6 +324,16 @@ func (p Proposal) IntoProto() *vegapb.Proposal {
 		Reason:       p.Reason,
 		ErrorDetails: p.ErrorDetails,
 	}
+
+	if p.Rationale != nil {
+		proposal.Rationale = &vegapb.ProposalRationale{
+			Description: p.Rationale.Description,
+			Hash:        p.Rationale.Hash,
+			Url:         p.Rationale.URL,
+		}
+	}
+
+	return proposal
 }
 
 func (v Vote) IntoProto() *vegapb.Vote {
@@ -414,6 +439,12 @@ func (n NewMarketCommitment) DeepClone() *NewMarketCommitment {
 	return cpy
 }
 
+type ProposalRationale struct {
+	Description string
+	Hash        string
+	URL         string
+}
+
 type ProposalTerms struct {
 	ClosingTimestamp    int64
 	EnactmentTimestamp  int64
@@ -477,15 +508,7 @@ type ProposalTermsNewAsset struct {
 	NewAsset *NewAsset
 }
 
-type NewFreeformDetails struct {
-	URL         string
-	Description string
-	Hash        string
-}
-
-type NewFreeform struct {
-	Changes *NewFreeformDetails
-}
+type NewFreeform struct{}
 
 type ProposalTermsNewFreeform struct {
 	NewFreeform *NewFreeform
@@ -650,6 +673,17 @@ func ProposalTermsFromProto(p *vegapb.ProposalTerms) *ProposalTerms {
 	}
 }
 
+func ProposalRationaleFromProto(p *vegapb.ProposalRationale) *ProposalRationale {
+	if p == nil {
+		return nil
+	}
+	return &ProposalRationale{
+		Description: p.Description,
+		Hash:        p.Hash,
+		URL:         p.Url,
+	}
+}
+
 func NewNewMarketFromProto(p *vegapb.ProposalTerms_NewMarket) *ProposalTermsNewMarket {
 	var newMarket *NewMarket
 	if p.NewMarket != nil {
@@ -700,20 +734,9 @@ func NewNewAssetFromProto(p *vegapb.ProposalTerms_NewAsset) *ProposalTermsNewAss
 	}
 }
 
-func NewNewFreeformFromProto(p *vegapb.ProposalTerms_NewFreeform) *ProposalTermsNewFreeform {
-	var newFreeform *NewFreeform
-	if p.NewFreeform != nil && p.NewFreeform.Changes != nil {
-		newFreeform = &NewFreeform{
-			Changes: &NewFreeformDetails{
-				URL:         p.NewFreeform.Changes.Url,
-				Description: p.NewFreeform.Changes.Description,
-				Hash:        p.NewFreeform.Changes.Hash,
-			},
-		}
-	}
-
+func NewNewFreeformFromProto(_ *vegapb.ProposalTerms_NewFreeform) *ProposalTermsNewFreeform {
 	return &ProposalTermsNewFreeform{
-		NewFreeform: newFreeform,
+		NewFreeform: &NewFreeform{},
 	}
 }
 
@@ -1154,13 +1177,7 @@ func (f ProposalTermsNewFreeform) DeepClone() proposalTerm {
 }
 
 func (n NewFreeform) IntoProto() *vegapb.NewFreeform {
-	return &vegapb.NewFreeform{
-		Changes: &vegapb.NewFreeformDetails{
-			Url:         n.Changes.URL,
-			Description: n.Changes.Description,
-			Hash:        n.Changes.Hash,
-		},
-	}
+	return &vegapb.NewFreeform{}
 }
 
 func (n NewFreeform) String() string {
@@ -1168,13 +1185,7 @@ func (n NewFreeform) String() string {
 }
 
 func (n NewFreeform) DeepClone() *NewFreeform {
-	return &NewFreeform{
-		Changes: &NewFreeformDetails{
-			URL:         n.Changes.URL,
-			Description: n.Changes.Description,
-			Hash:        n.Changes.Hash,
-		},
-	}
+	return &NewFreeform{}
 }
 
 func UpdateMarketFromProto(p *vegapb.ProposalTerms_UpdateMarket) *ProposalTermsUpdateMarket {